+++ history.md
@@ -8,3 +8,4 @@
   - *fixed:* Badly identified declaring types (in certain circumstances)
 - 1.1.2:
   - *fixed:* Misinterpreted "ref" modifier (in certain circumstances)
+  - *fixed:* Proxy generation failed on interface methods having parameter names that were used internally